FEATURE: moves selectBoards
