+++ libs/spirit-islander/util/src/index.ts
@@ -23,5 +23,5 @@ export * from './lib/game/get-adversary-by-id';
 export * from './lib/game/get-difficulty';
 export * from './lib/game/get-options';
 export * from './lib/functions/get-valid-combos';
-export * from './lib/functions/select-boards';
+export * from './lib/app/select-boards';
 export * from './lib/functions/update-model';
